+++ src/BsssYaml.hh
@@ -36,7 +36,7 @@ namespace Bsss {
             void setEdefEnable(int chn, uint32_t enable);
 
             // channel number correspnds to the BSSS data channels
-            void setChannelSevr(int chn, uint32_t sevr);
+            void setChannelSevr(int chn, uint64_t sevr);
             void setChannelMask(int chn, uint32_t enable);
             void setPacketSize(uint32_t size);
             void enablePacket(uint32_t enable);
